html/rendering/non-replaced-elements/the-fieldset-and-legend-elements/fieldset-percentage-block-size.html

<!doctype html>
<title>fieldset percentage block size</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<style>
 fieldset { block-size: 100px; }
 legend { block-size: 100%; background: lime; }
 fieldset > div { block-size: 100%; background: green; font-weight: bold; }
</style>
<div style="writing-mode: horizontal-tb">
 <fieldset>
  <legend>rendered legend</legend>
  <legend>second legend</legend>
  <div>div</div>
 </fieldset>
</div>

<div style="writing-mode: vertical-lr">
 <fieldset>
  <legend>rendered legend</legend>
  <legend>second legend</legend>
  <div>div</div>
 </fieldset>
</div>

<div style="writing-mode: vertical-rl">
 <fieldset>
  <legend>rendered legend</legend>
  <legend>second legend</legend>
  <div>div</div>
 </fieldset>
</div>

<script>
 for (const div of document.querySelectorAll('div[style]')) {
   for (const el of div.firstElementChild.children) {
     test(() => {
       assert_equals(getComputedStyle(el).blockSize, '100px');
     }, `${el.textContent} (${div.getAttribute('style')})`);
   }
 }
</script>
